Move PolicyMixture to its own file
